skills/sandbox/pyproject.toml
[project]
name = "haiku-skills-sandbox"
version = "0.13.0"
description = "Docker sandbox skill for haiku.skills"
authors = [{ name = "Yiorgis Gozadinos", email = "ggozadinos@gmail.com" }]
license = { text = "MIT" }
readme = { file = "README.md", content-type = "text/markdown" }
requires-python = ">=3.12"
keywords = ["haiku-skills", "agent-skills", "pydantic-ai", "sandbox", "docker"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Typing :: Typed",
]
dependencies = ["haiku.skills>=0.13.0", "pydantic-ai-backend[docker,console]>=0.1.9"]

[project.urls]
Homepage = "https://github.com/ggozad/haiku.skills"
Repository = "https://github.com/ggozad/haiku.skills"
Issues = "https://github.com/ggozad/haiku.skills/issues"

[project.entry-points."haiku.skills"]
sandbox = "haiku_skills_sandbox:create_skill"

[tool.uv.sources]
haiku-skills = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
